CR: split non strict tests into separate file, move helper method
