+++ engine/runtime/src/test/java/org/enso/interpreter/test/TestBase.java
@@ -32,14 +32,6 @@ public abstract class TestBase {
     return context;
   }
 
-  protected static Context createNonStrictContext(OutputStream out) {
-    var context =
-        defaultContextBuilder().out(out).option(RuntimeOptions.STRICT_ERRORS, "false").build();
-    final Map<String, Language> langs = context.getEngine().getLanguages();
-    assertNotNull("Enso found: " + langs, langs.get("enso"));
-    return context;
-  }
-
   protected static Context createDefaultContext(OutputStream out) {
     var context = defaultContextBuilder().out(out).build();
     final Map<String, Language> langs = context.getEngine().getLanguages();
